Refactor src/client/client.test.ts (993 lines) into per-area sibling test files (warren-5b69 / pl-9088 step 11)
The monolithic client.test.ts had a single 655-line describe block
('WarrenClient.dispatch + waitForRun') that triggered Biome's
noExcessiveLinesPerFunction rule and forced an override entry. Split
into seven focused test files alongside the original, sharing
jsonResponse/stub helpers via src/client/test-helpers.ts:

  client.test.ts                  121  (constructor, fromEnv, getProject, createRun, error rehydration)
  client.projects-agents.test.ts  168  (listProjects, createProject, refreshProject, *Agents)
  client.probe.test.ts             41  (probe success/timeout/unreachable)
  client.runs.test.ts             232  (dispatch, steer, getRun, waitForRun)
  client.stream.test.ts           177  (streamRunEvents NDJSON parsing + error paths)
  client.plots.test.ts            195  (list/get/create/editIntent/changeStatus/sync)
  client.plan-runs.test.ts         69  (createPlanRun, getPlanRun, listPlanRuns)

All files now <=500 lines; all 52 tests still pass.

- Removed the noExcessiveLinesPerFunction override entry for
  src/client/client.test.ts from biome.json (refresh.test.ts entry
  retained for warren-a0e8).
- Dropped the file-size-budget entry for src/client/client.test.ts
  from scripts/file-size-budgets.json (ratchet only goes down).

bun run check:all green.
